refactor: REGISTERED가 아닌 이미지 조회 메서드 수정
