+++ be/src/main/java/kr/codesquad/jazzmeet/image/repository/ImageRepository.java
@@ -1,6 +1,5 @@
 package kr.codesquad.jazzmeet.image.repository;
 
-import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -13,6 +12,4 @@ import kr.codesquad.jazzmeet.image.entity.ImageStatus;
 public interface ImageRepository extends JpaRepository<Image, Long> {
 
 	Optional<Image> findByIdAndStatusNot(Long imageId, ImageStatus status);
-
-	List<Image> findAllByStatusNot(ImageStatus status);
 }
